product & order controller

+++ src/main/java/com/nk/clothify_backend/repository/ProductRepository.java
@@ -19,8 +19,8 @@ public interface ProductRepository extends JpaRepository<ProductEntity,Long> {
                     "CASE WHEN :sort = 'price_high' THEN p.discountedPrice END DESC"
     )
     public List<ProductEntity> filterProducts(@Param("category") String category,
-                                              @Param("minPrice") Double minPrice,
-                                              @Param("maxPrice") Double maxPrice,
+                                              @Param("minPrice") Integer minPrice,
+                                              @Param("maxPrice") Integer maxPrice,
                                               @Param("minDiscount") Integer minDiscount,
                                               @Param("sort") String sort);
