+++ src/main/java/com/nk/clothify_backend/service/ProductService.java
@@ -14,7 +14,7 @@ public interface ProductService {
     public Product updateProduct(Long productId,Product req) throws ProductException;
     public Product findProductById(Long productId) throws ProductException;
     public List<Product> findProductByCategory(String category);
-    public Page<Product> getAllProduct(String category, List<String> colors, List<String> sizes, Double minPrice, Double maxPrice,
+    public Page<Product> getAllProduct(String category, List<String> colors, List<String> sizes, Integer minPrice, Integer maxPrice,
                                        Integer minDiscount, String sort, String stock, Integer pageNumber, Integer pageSize);
 
 }
